numeric addition
